Add files via upload
The above file will automate the whole WPS and WRF process from setting the domain in the namelist.wps to submitting a job in HPC. User has to use a '.sh' script with name 'parallel.sh'. this script should include both the commands to execute real and wrf. This python file should be stored in WPS folder. 
The layout of folder structure should be such that there should be 3 folders at the same level namely; 1.data (to store meteorological data), 2.WPS and 3.WRF.
Further physics schemes could be set by opening the python code. We welcome you to use and make and value addition to the code.